+++ spec/spec_helper.rb
@@ -1,13 +1,24 @@
 # frozen_string_literal: true
 
 require "apify"
+require "webmock/rspec"
+
+Dir[File.join(__dir__, "support", "**", "*.rb")].each { |file| require file }
+
+WebMock.disable_net_connect!(allow_localhost: true)
 
 RSpec.configure do |config|
-  # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = ".rspec_status"
-
-  # Disable RSpec exposing methods globally on `Module` and `main`
   config.disable_monkey_patching!
+
+  config.before do
+    Apify.reset!
+    Apify.configure do |apify_config|
+      apify_config.api_token = "test-apify-token"
+      apify_config.max_retries = 0
+      apify_config.sleep_fn = ->(_seconds) {}
+    end
+  end
 
   config.expect_with :rspec do |c|
     c.syntax = :expect